pyproject.toml
[project]
name = "egghunt-event"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "waitress>=3.0.2",
]
